Convert List Item Selector to interface and introduce Single Item Selector
